Move AndroidEvent builders to builder class
